WIP: Extract common, Timeline-related utils to their own module.
