+++ src/web/ui/utils/timeline/TimelineGraphBg.tsx
@@ -2,7 +2,7 @@ import { MeterEntry, SensorEntry } from 'core/models/model';
 import { css } from 'emotion';
 import React from 'react';
 import { isNotNull } from 'server/utils/types';
-import { bgToTop, ExtendedTimelineConfig, tsToLeft } from 'web/ui/utils/timeline/Timeline';
+import { bgToTop, ExtendedTimelineConfig, tsToLeft } from 'web/ui/utils/timeline/utils';
 
 type Props = {
   timelineConfig: ExtendedTimelineConfig;
