+++ src/web/ui/utils/timeline/TimelineScaleTs.tsx
@@ -3,7 +3,7 @@ import { css } from 'emotion';
 import { range } from 'lodash';
 import { DateTime } from 'luxon';
 import React, { Fragment } from 'react';
-import { ExtendedTimelineConfig, tsToLeft } from 'web/ui/utils/timeline/Timeline';
+import { ExtendedTimelineConfig, tsToLeft } from 'web/ui/utils/timeline/utils';
 
 type Props = {
   timelineConfig: ExtendedTimelineConfig;
